arcade/requirements.txt
python-dotenv
asyncio
python-socketio[asyncio_client]
pyjwt
evdev
